Fixed tests. Added more tests for `memcopy` and fixed a bug. Added missed tests for message serialization. Removed unused unsafe code. Removed `RESERVED**` variants from the `PanicReason` enum.
