+++ fuel-asm/src/panic_reason.rs
@@ -81,444 +81,8 @@ pub enum PanicReason {
     ContractMismatch = 0x23,
     /// No more nested calls are allowed.
     NestedCallLimitReached = 0x24,
-    /// RESERV25
-    RESERV25 = 0x25,
-    /// RESERV26
-    RESERV26 = 0x26,
-    /// RESERV27
-    RESERV27 = 0x27,
-    /// RESERV28
-    RESERV28 = 0x28,
-    /// RESERV29
-    RESERV29 = 0x29,
-    /// RESERV2A
-    RESERV2A = 0x2a,
-    /// RESERV2B
-    RESERV2B = 0x2b,
-    /// RESERV2C
-    RESERV2C = 0x2c,
-    /// RESERV2D
-    RESERV2D = 0x2d,
-    /// RESERV2E
-    RESERV2E = 0x2e,
-    /// RESERV2F
-    RESERV2F = 0x2f,
-    /// RESERV30
-    RESERV30 = 0x30,
-    /// RESERV31
-    RESERV31 = 0x31,
-    /// RESERV32
-    RESERV32 = 0x32,
-    /// RESERV33
-    RESERV33 = 0x33,
-    /// RESERV34
-    RESERV34 = 0x34,
-    /// RESERV35
-    RESERV35 = 0x35,
-    /// RESERV36
-    RESERV36 = 0x36,
-    /// RESERV37
-    RESERV37 = 0x37,
-    /// RESERV38
-    RESERV38 = 0x38,
-    /// RESERV39
-    RESERV39 = 0x39,
-    /// RESERV3A
-    RESERV3A = 0x3a,
-    /// RESERV3B
-    RESERV3B = 0x3b,
-    /// RESERV3C
-    RESERV3C = 0x3c,
-    /// RESERV3D
-    RESERV3D = 0x3d,
-    /// RESERV3E
-    RESERV3E = 0x3e,
-    /// RESERV3F
-    RESERV3F = 0x3f,
-    /// RESERV40
-    RESERV40 = 0x40,
-    /// RESERV41
-    RESERV41 = 0x41,
-    /// RESERV42
-    RESERV42 = 0x42,
-    /// RESERV43
-    RESERV43 = 0x43,
-    /// RESERV44
-    RESERV44 = 0x44,
-    /// RESERV45
-    RESERV45 = 0x45,
-    /// RESERV46
-    RESERV46 = 0x46,
-    /// RESERV47
-    RESERV47 = 0x47,
-    /// RESERV48
-    RESERV48 = 0x48,
-    /// RESERV49
-    RESERV49 = 0x49,
-    /// RESERV4A
-    RESERV4A = 0x4a,
-    /// RESERV4B
-    RESERV4B = 0x4b,
-    /// RESERV4C
-    RESERV4C = 0x4c,
-    /// RESERV4D
-    RESERV4D = 0x4d,
-    /// RESERV4E
-    RESERV4E = 0x4e,
-    /// RESERV4F
-    RESERV4F = 0x4f,
-    /// RESERV50
-    RESERV50 = 0x50,
-    /// RESERV51
-    RESERV51 = 0x51,
-    /// RESERV52
-    RESERV52 = 0x52,
-    /// RESERV53
-    RESERV53 = 0x53,
-    /// RESERV54
-    RESERV54 = 0x54,
-    /// RESERV55
-    RESERV55 = 0x55,
-    /// RESERV56
-    RESERV56 = 0x56,
-    /// RESERV57
-    RESERV57 = 0x57,
-    /// RESERV58
-    RESERV58 = 0x58,
-    /// RESERV59
-    RESERV59 = 0x59,
-    /// RESERV5A
-    RESERV5A = 0x5a,
-    /// RESERV5B
-    RESERV5B = 0x5b,
-    /// RESERV5C
-    RESERV5C = 0x5c,
-    /// RESERV5D
-    RESERV5D = 0x5d,
-    /// RESERV5E
-    RESERV5E = 0x5e,
-    /// RESERV5F
-    RESERV5F = 0x5f,
-    /// RESERV60
-    RESERV60 = 0x60,
-    /// RESERV61
-    RESERV61 = 0x61,
-    /// RESERV62
-    RESERV62 = 0x62,
-    /// RESERV63
-    RESERV63 = 0x63,
-    /// RESERV64
-    RESERV64 = 0x64,
-    /// RESERV65
-    RESERV65 = 0x65,
-    /// RESERV66
-    RESERV66 = 0x66,
-    /// RESERV67
-    RESERV67 = 0x67,
-    /// RESERV68
-    RESERV68 = 0x68,
-    /// RESERV69
-    RESERV69 = 0x69,
-    /// RESERV6A
-    RESERV6A = 0x6a,
-    /// RESERV6B
-    RESERV6B = 0x6b,
-    /// RESERV6C
-    RESERV6C = 0x6c,
-    /// RESERV6D
-    RESERV6D = 0x6d,
-    /// RESERV6E
-    RESERV6E = 0x6e,
-    /// RESERV6F
-    RESERV6F = 0x6f,
-    /// RESERV70
-    RESERV70 = 0x70,
-    /// RESERV71
-    RESERV71 = 0x71,
-    /// RESERV72
-    RESERV72 = 0x72,
-    /// RESERV73
-    RESERV73 = 0x73,
-    /// RESERV74
-    RESERV74 = 0x74,
-    /// RESERV75
-    RESERV75 = 0x75,
-    /// RESERV76
-    RESERV76 = 0x76,
-    /// RESERV77
-    RESERV77 = 0x77,
-    /// RESERV78
-    RESERV78 = 0x78,
-    /// RESERV79
-    RESERV79 = 0x79,
-    /// RESERV7A
-    RESERV7A = 0x7a,
-    /// RESERV7B
-    RESERV7B = 0x7b,
-    /// RESERV7C
-    RESERV7C = 0x7c,
-    /// RESERV7D
-    RESERV7D = 0x7d,
-    /// RESERV7E
-    RESERV7E = 0x7e,
-    /// RESERV7F
-    RESERV7F = 0x7f,
-    /// RESERV80
-    RESERV80 = 0x80,
-    /// RESERV81
-    RESERV81 = 0x81,
-    /// RESERV82
-    RESERV82 = 0x82,
-    /// RESERV83
-    RESERV83 = 0x83,
-    /// RESERV84
-    RESERV84 = 0x84,
-    /// RESERV85
-    RESERV85 = 0x85,
-    /// RESERV86
-    RESERV86 = 0x86,
-    /// RESERV87
-    RESERV87 = 0x87,
-    /// RESERV88
-    RESERV88 = 0x88,
-    /// RESERV89
-    RESERV89 = 0x89,
-    /// RESERV8A
-    RESERV8A = 0x8a,
-    /// RESERV8B
-    RESERV8B = 0x8b,
-    /// RESERV8C
-    RESERV8C = 0x8c,
-    /// RESERV8D
-    RESERV8D = 0x8d,
-    /// RESERV8E
-    RESERV8E = 0x8e,
-    /// RESERV8F
-    RESERV8F = 0x8f,
-    /// RESERV90
-    RESERV90 = 0x90,
-    /// RESERV91
-    RESERV91 = 0x91,
-    /// RESERV92
-    RESERV92 = 0x92,
-    /// RESERV93
-    RESERV93 = 0x93,
-    /// RESERV94
-    RESERV94 = 0x94,
-    /// RESERV95
-    RESERV95 = 0x95,
-    /// RESERV96
-    RESERV96 = 0x96,
-    /// RESERV97
-    RESERV97 = 0x97,
-    /// RESERV98
-    RESERV98 = 0x98,
-    /// RESERV99
-    RESERV99 = 0x99,
-    /// RESERV9A
-    RESERV9A = 0x9a,
-    /// RESERV9B
-    RESERV9B = 0x9b,
-    /// RESERV9C
-    RESERV9C = 0x9c,
-    /// RESERV9D
-    RESERV9D = 0x9d,
-    /// RESERV9E
-    RESERV9E = 0x9e,
-    /// RESERV9F
-    RESERV9F = 0x9f,
-    /// RESERVA0
-    RESERVA0 = 0xa0,
-    /// RESERVA1
-    RESERVA1 = 0xa1,
-    /// RESERVA2
-    RESERVA2 = 0xa2,
-    /// RESERVA3
-    RESERVA3 = 0xa3,
-    /// RESERVA4
-    RESERVA4 = 0xa4,
-    /// RESERVA5
-    RESERVA5 = 0xa5,
-    /// RESERVA6
-    RESERVA6 = 0xa6,
-    /// RESERVA7
-    RESERVA7 = 0xa7,
-    /// RESERVA8
-    RESERVA8 = 0xa8,
-    /// RESERVA9
-    RESERVA9 = 0xa9,
-    /// RESERVAA
-    RESERVAA = 0xaa,
-    /// RESERVAB
-    RESERVAB = 0xab,
-    /// RESERVAC
-    RESERVAC = 0xac,
-    /// RESERVAD
-    RESERVAD = 0xad,
-    /// RESERVAE
-    RESERVAE = 0xae,
-    /// RESERVAF
-    RESERVAF = 0xaf,
-    /// RESERVB0
-    RESERVB0 = 0xb0,
-    /// RESERVB1
-    RESERVB1 = 0xb1,
-    /// RESERVB2
-    RESERVB2 = 0xb2,
-    /// RESERVB3
-    RESERVB3 = 0xb3,
-    /// RESERVB4
-    RESERVB4 = 0xb4,
-    /// RESERVB5
-    RESERVB5 = 0xb5,
-    /// RESERVB6
-    RESERVB6 = 0xb6,
-    /// RESERVB7
-    RESERVB7 = 0xb7,
-    /// RESERVB8
-    RESERVB8 = 0xb8,
-    /// RESERVB9
-    RESERVB9 = 0xb9,
-    /// RESERVBA
-    RESERVBA = 0xba,
-    /// RESERVBB
-    RESERVBB = 0xbb,
-    /// RESERVBC
-    RESERVBC = 0xbc,
-    /// RESERVBD
-    RESERVBD = 0xbd,
-    /// RESERVBE
-    RESERVBE = 0xbe,
-    /// RESERVBF
-    RESERVBF = 0xbf,
-    /// RESERVC0
-    RESERVC0 = 0xc0,
-    /// RESERVC1
-    RESERVC1 = 0xc1,
-    /// RESERVC2
-    RESERVC2 = 0xc2,
-    /// RESERVC3
-    RESERVC3 = 0xc3,
-    /// RESERVC4
-    RESERVC4 = 0xc4,
-    /// RESERVC5
-    RESERVC5 = 0xc5,
-    /// RESERVC6
-    RESERVC6 = 0xc6,
-    /// RESERVC7
-    RESERVC7 = 0xc7,
-    /// RESERVC8
-    RESERVC8 = 0xc8,
-    /// RESERVC9
-    RESERVC9 = 0xc9,
-    /// RESERVCA
-    RESERVCA = 0xca,
-    /// RESERVCB
-    RESERVCB = 0xcb,
-    /// RESERVCC
-    RESERVCC = 0xcc,
-    /// RESERVCD
-    RESERVCD = 0xcd,
-    /// RESERVCE
-    RESERVCE = 0xce,
-    /// RESERVCF
-    RESERVCF = 0xcf,
-    /// RESERVD0
-    RESERVD0 = 0xd0,
-    /// RESERVD1
-    RESERVD1 = 0xd1,
-    /// RESERVD2
-    RESERVD2 = 0xd2,
-    /// RESERVD3
-    RESERVD3 = 0xd3,
-    /// RESERVD4
-    RESERVD4 = 0xd4,
-    /// RESERVD5
-    RESERVD5 = 0xd5,
-    /// RESERVD6
-    RESERVD6 = 0xd6,
-    /// RESERVD7
-    RESERVD7 = 0xd7,
-    /// RESERVD8
-    RESERVD8 = 0xd8,
-    /// RESERVD9
-    RESERVD9 = 0xd9,
-    /// RESERVDA
-    RESERVDA = 0xda,
-    /// RESERVDB
-    RESERVDB = 0xdb,
-    /// RESERVDC
-    RESERVDC = 0xdc,
-    /// RESERVDD
-    RESERVDD = 0xdd,
-    /// RESERVDE
-    RESERVDE = 0xde,
-    /// RESERVDF
-    RESERVDF = 0xdf,
-    /// RESERVE0
-    RESERVE0 = 0xe0,
-    /// RESERVE1
-    RESERVE1 = 0xe1,
-    /// RESERVE2
-    RESERVE2 = 0xe2,
-    /// RESERVE3
-    RESERVE3 = 0xe3,
-    /// RESERVE4
-    RESERVE4 = 0xe4,
-    /// RESERVE5
-    RESERVE5 = 0xe5,
-    /// RESERVE6
-    RESERVE6 = 0xe6,
-    /// RESERVE7
-    RESERVE7 = 0xe7,
-    /// RESERVE8
-    RESERVE8 = 0xe8,
-    /// RESERVE9
-    RESERVE9 = 0xe9,
-    /// RESERVEA
-    RESERVEA = 0xea,
-    /// RESERVEB
-    RESERVEB = 0xeb,
-    /// RESERVEC
-    RESERVEC = 0xec,
-    /// RESERVED
-    RESERVED = 0xed,
-    /// RESERVEE
-    RESERVEE = 0xee,
-    /// RESERVEF
-    RESERVEF = 0xef,
-    /// RESERVF0
-    RESERVF0 = 0xf0,
-    /// RESERVF1
-    RESERVF1 = 0xf1,
-    /// RESERVF2
-    RESERVF2 = 0xf2,
-    /// RESERVF3
-    RESERVF3 = 0xf3,
-    /// RESERVF4
-    RESERVF4 = 0xf4,
-    /// RESERVF5
-    RESERVF5 = 0xf5,
-    /// RESERVF6
-    RESERVF6 = 0xf6,
-    /// RESERVF7
-    RESERVF7 = 0xf7,
-    /// RESERVF8
-    RESERVF8 = 0xf8,
-    /// RESERVF9
-    RESERVF9 = 0xf9,
-    /// RESERVFA
-    RESERVFA = 0xfa,
-    /// RESERVFB
-    RESERVFB = 0xfb,
-    /// RESERVFC
-    RESERVFC = 0xfc,
-    /// RESERVFD
-    RESERVFD = 0xfd,
-    /// RESERVFE
-    RESERVFE = 0xfe,
-    /// RESERVFF
-    RESERVFF = 0xff,
+    /// The byte can't be mapped to any known `PanicReason`.
+    UnknownPanicReason = 0x25,
 }
 
 impl fmt::Display for PanicReason {
@@ -541,8 +105,6 @@ impl From<convert::Infallible> for PanicReason {
     }
 }
 
-// TODO: `PanicReason` shouldn't list reserved variants, and this should be `TryFrom` with no
-// `unsafe`.
 impl From<u8> for PanicReason {
     fn from(b: u8) -> Self {
         use PanicReason::*;
@@ -584,225 +146,7 @@ impl From<u8> for PanicReason {
             0x22 => ContractIdAlreadyDeployed,
             0x23 => ContractMismatch,
             0x24 => NestedCallLimitReached,
-            0x25 => RESERV25,
-            0x26 => RESERV26,
-            0x27 => RESERV27,
-            0x28 => RESERV28,
-            0x29 => RESERV29,
-            0x2a => RESERV2A,
-            0x2b => RESERV2B,
-            0x2c => RESERV2C,
-            0x2d => RESERV2D,
-            0x2e => RESERV2E,
-            0x2f => RESERV2F,
-            0x30 => RESERV30,
-            0x31 => RESERV31,
-            0x32 => RESERV32,
-            0x33 => RESERV33,
-            0x34 => RESERV34,
-            0x35 => RESERV35,
-            0x36 => RESERV36,
-            0x37 => RESERV37,
-            0x38 => RESERV38,
-            0x39 => RESERV39,
-            0x3a => RESERV3A,
-            0x3b => RESERV3B,
-            0x3c => RESERV3C,
-            0x3d => RESERV3D,
-            0x3e => RESERV3E,
-            0x3f => RESERV3F,
-            0x40 => RESERV40,
-            0x41 => RESERV41,
-            0x42 => RESERV42,
-            0x43 => RESERV43,
-            0x44 => RESERV44,
-            0x45 => RESERV45,
-            0x46 => RESERV46,
-            0x47 => RESERV47,
-            0x48 => RESERV48,
-            0x49 => RESERV49,
-            0x4a => RESERV4A,
-            0x4b => RESERV4B,
-            0x4c => RESERV4C,
-            0x4d => RESERV4D,
-            0x4e => RESERV4E,
-            0x4f => RESERV4F,
-            0x50 => RESERV50,
-            0x51 => RESERV51,
-            0x52 => RESERV52,
-            0x53 => RESERV53,
-            0x54 => RESERV54,
-            0x55 => RESERV55,
-            0x56 => RESERV56,
-            0x57 => RESERV57,
-            0x58 => RESERV58,
-            0x59 => RESERV59,
-            0x5a => RESERV5A,
-            0x5b => RESERV5B,
-            0x5c => RESERV5C,
-            0x5d => RESERV5D,
-            0x5e => RESERV5E,
-            0x5f => RESERV5F,
-            0x60 => RESERV60,
-            0x61 => RESERV61,
-            0x62 => RESERV62,
-            0x63 => RESERV63,
-            0x64 => RESERV64,
-            0x65 => RESERV65,
-            0x66 => RESERV66,
-            0x67 => RESERV67,
-            0x68 => RESERV68,
-            0x69 => RESERV69,
-            0x6a => RESERV6A,
-            0x6b => RESERV6B,
-            0x6c => RESERV6C,
-            0x6d => RESERV6D,
-            0x6e => RESERV6E,
-            0x6f => RESERV6F,
-            0x70 => RESERV70,
-            0x71 => RESERV71,
-            0x72 => RESERV72,
-            0x73 => RESERV73,
-            0x74 => RESERV74,
-            0x75 => RESERV75,
-            0x76 => RESERV76,
-            0x77 => RESERV77,
-            0x78 => RESERV78,
-            0x79 => RESERV79,
-            0x7a => RESERV7A,
-            0x7b => RESERV7B,
-            0x7c => RESERV7C,
-            0x7d => RESERV7D,
-            0x7e => RESERV7E,
-            0x7f => RESERV7F,
-            0x80 => RESERV80,
-            0x81 => RESERV81,
-            0x82 => RESERV82,
-            0x83 => RESERV83,
-            0x84 => RESERV84,
-            0x85 => RESERV85,
-            0x86 => RESERV86,
-            0x87 => RESERV87,
-            0x88 => RESERV88,
-            0x89 => RESERV89,
-            0x8a => RESERV8A,
-            0x8b => RESERV8B,
-            0x8c => RESERV8C,
-            0x8d => RESERV8D,
-            0x8e => RESERV8E,
-            0x8f => RESERV8F,
-            0x90 => RESERV90,
-            0x91 => RESERV91,
-            0x92 => RESERV92,
-            0x93 => RESERV93,
-            0x94 => RESERV94,
-            0x95 => RESERV95,
-            0x96 => RESERV96,
-            0x97 => RESERV97,
-            0x98 => RESERV98,
-            0x99 => RESERV99,
-            0x9a => RESERV9A,
-            0x9b => RESERV9B,
-            0x9c => RESERV9C,
-            0x9d => RESERV9D,
-            0x9e => RESERV9E,
-            0x9f => RESERV9F,
-            0xa0 => RESERVA0,
-            0xa1 => RESERVA1,
-            0xa2 => RESERVA2,
-            0xa3 => RESERVA3,
-            0xa4 => RESERVA4,
-            0xa5 => RESERVA5,
-            0xa6 => RESERVA6,
-            0xa7 => RESERVA7,
-            0xa8 => RESERVA8,
-            0xa9 => RESERVA9,
-            0xaa => RESERVAA,
-            0xab => RESERVAB,
-            0xac => RESERVAC,
-            0xad => RESERVAD,
-            0xae => RESERVAE,
-            0xaf => RESERVAF,
-            0xb0 => RESERVB0,
-            0xb1 => RESERVB1,
-            0xb2 => RESERVB2,
-            0xb3 => RESERVB3,
-            0xb4 => RESERVB4,
-            0xb5 => RESERVB5,
-            0xb6 => RESERVB6,
-            0xb7 => RESERVB7,
-            0xb8 => RESERVB8,
-            0xb9 => RESERVB9,
-            0xba => RESERVBA,
-            0xbb => RESERVBB,
-            0xbc => RESERVBC,
-            0xbd => RESERVBD,
-            0xbe => RESERVBE,
-            0xbf => RESERVBF,
-            0xc0 => RESERVC0,
-            0xc1 => RESERVC1,
-            0xc2 => RESERVC2,
-            0xc3 => RESERVC3,
-            0xc4 => RESERVC4,
-            0xc5 => RESERVC5,
-            0xc6 => RESERVC6,
-            0xc7 => RESERVC7,
-            0xc8 => RESERVC8,
-            0xc9 => RESERVC9,
-            0xca => RESERVCA,
-            0xcb => RESERVCB,
-            0xcc => RESERVCC,
-            0xcd => RESERVCD,
-            0xce => RESERVCE,
-            0xcf => RESERVCF,
-            0xd0 => RESERVD0,
-            0xd1 => RESERVD1,
-            0xd2 => RESERVD2,
-            0xd3 => RESERVD3,
-            0xd4 => RESERVD4,
-            0xd5 => RESERVD5,
-            0xd6 => RESERVD6,
-            0xd7 => RESERVD7,
-            0xd8 => RESERVD8,
-            0xd9 => RESERVD9,
-            0xda => RESERVDA,
-            0xdb => RESERVDB,
-            0xdc => RESERVDC,
-            0xdd => RESERVDD,
-            0xde => RESERVDE,
-            0xdf => RESERVDF,
-            0xe0 => RESERVE0,
-            0xe1 => RESERVE1,
-            0xe2 => RESERVE2,
-            0xe3 => RESERVE3,
-            0xe4 => RESERVE4,
-            0xe5 => RESERVE5,
-            0xe6 => RESERVE6,
-            0xe7 => RESERVE7,
-            0xe8 => RESERVE8,
-            0xe9 => RESERVE9,
-            0xea => RESERVEA,
-            0xeb => RESERVEB,
-            0xec => RESERVEC,
-            0xed => RESERVED,
-            0xee => RESERVEE,
-            0xef => RESERVEF,
-            0xf0 => RESERVF0,
-            0xf1 => RESERVF1,
-            0xf2 => RESERVF2,
-            0xf3 => RESERVF3,
-            0xf4 => RESERVF4,
-            0xf5 => RESERVF5,
-            0xf6 => RESERVF6,
-            0xf7 => RESERVF7,
-            0xf8 => RESERVF8,
-            0xf9 => RESERVF9,
-            0xfa => RESERVFA,
-            0xfb => RESERVFB,
-            0xfc => RESERVFC,
-            0xfd => RESERVFD,
-            0xfe => RESERVFE,
-            0xff => RESERVFF,
+            _ => UnknownPanicReason,
         }
     }
 }
@@ -828,10 +172,15 @@ mod tests {
 
     #[test]
     fn test_u8_panic_reason_round_trip() {
-        for i in 0..=255 {
+        for i in 0..0x25 {
             let reason = PanicReason::from(i);
             let i2 = reason as u8;
             assert_eq!(i, i2);
+        }
+        for i in 0x25..=255 {
+            let reason = PanicReason::from(i);
+            let i2 = reason as u8;
+            assert_eq!(0x25, i2);
         }
     }
 }
