+++ fuel-tx/src/receipt/receipt_std.rs
@@ -4,13 +4,11 @@ use fuel_asm::InstructionResult;
 use fuel_types::bytes::{self, SizedBytes, WORD_SIZE};
 use fuel_types::{MemLayout, MemLocType, Word};
 
-use crate::receipt::receipt_std::sizes::CallSizes;
 use crate::receipt::script_result::ScriptExecutionResult;
+use crate::receipt::sizes::CallSizes;
 use std::io::{self, Write};
 
-use sizes::*;
-
-mod sizes;
+use crate::receipt::sizes::*;
 
 impl io::Read for Receipt {
     fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
@@ -347,11 +345,7 @@ impl io::Write for Receipt {
                 let pc = bytes::restore_word_at(buf, S::layout(S::LAYOUT.pc));
                 let is = bytes::restore_word_at(buf, S::layout(S::LAYOUT.is));
 
-                let (_, data, buf) = bytes::restore_bytes(full_buf.get(S::LEN..).ok_or(bytes::eof())?)?;
-
-                if !buf.is_empty() {
-                    return Err(bytes::eof());
-                }
+                let (_, data, _) = bytes::restore_bytes(full_buf.get(S::LEN..).ok_or(bytes::eof())?)?;
 
                 let id = id.into();
                 let digest = digest.into();
@@ -429,11 +423,7 @@ impl io::Write for Receipt {
                 let pc = bytes::restore_word_at(buf, S::layout(S::LAYOUT.pc));
                 let is = bytes::restore_word_at(buf, S::layout(S::LAYOUT.is));
 
-                let (_, data, buf) = bytes::restore_bytes(full_buf.get(S::LEN..).ok_or(bytes::eof())?)?;
-
-                if !buf.is_empty() {
-                    return Err(bytes::eof());
-                }
+                let (_, data, _) = bytes::restore_bytes(full_buf.get(S::LEN..).ok_or(bytes::eof())?)?;
 
                 let id = id.into();
                 let digest = digest.into();
@@ -512,11 +502,7 @@ impl io::Write for Receipt {
                 let len = bytes::restore_word_at(buf, S::layout(S::LAYOUT.len));
                 let digest = bytes::restore_at(buf, S::layout(S::LAYOUT.digest));
 
-                let (_, data, buf) = bytes::restore_bytes(full_buf.get(S::LEN..).ok_or(bytes::eof())?)?;
-
-                if !buf.is_empty() {
-                    return Err(bytes::eof());
-                }
+                let (_, data, _) = bytes::restore_bytes(full_buf.get(S::LEN..).ok_or(bytes::eof())?)?;
 
                 let message_id = message_id.into();
                 let sender = sender.into();
@@ -528,7 +514,8 @@ impl io::Write for Receipt {
             }
         }
 
-        Ok(full_buf.len() + WORD_SIZE)
+        let n = self.serialized_size();
+        Ok(n)
     }
 
     fn flush(&mut self) -> io::Result<()> {
