+++ fuel-types/src/types.rs
@@ -85,6 +85,7 @@ macro_rules! key_methods {
             }
 
             #[cfg(feature = "unsafe")]
+            #[allow(unsafe_code)]
             /// Add a conversion from arbitrary slices into owned
             ///
             /// # Safety
